news_sentiment_analyser/requirements.txt
openai
python-dotenv
pandas
psycopg2